requirements.txt
dlib==19.24.0
matplotlib==3.1.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.0.3
pandas==1.0.1
Pillow==9.4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
timm==0.4.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.42.1
transformers==4.30.2
Transplant==0.8.11
xlrd==1.2.0